// tsSrc/configs/ncov_china/enums.ts
export enum ChartName {
  CH_ADD_HISTORY = 'chinaAddHistoryData',
  CH_NOW_HISTORY = 'chinaNowHistoryData',
  CH_TOTAL_HISTORY = 'chinaTotalHistoryData',
  CH_DAY_MODIFY = 'chinaDayModify',
  GD_MODIFY = 'guangdongModify',
  GD_TOTAL_HISTORY = 'guangdongTotalHistory',
  GZ_MODIFY_HISTORY = 'guangzhouModifyHistory'
}